Add PR374: More interaction tgui updates
