+++ tgui/packages/tgui/interfaces/MobInteraction/tabs/index.tsx
@@ -1,0 +1,4 @@
+export { InteractionsTab } from './InteractionsTab';
+export { GenitalTab } from './GenitalTab';
+export { CharacterPrefsTab } from './CharacterPrefsTab';
+export { ContentPreferencesTab } from './ContentPreferencesTab';
